refactor: (#56) useAuthStore 위치 변경
