+++ src/entities/user/index.ts
@@ -1,1 +1,2 @@
-export * from './model';
+export * from './model/types';
+export { useAuthStore } from './model/useAuthStore';
